fix: change amount type from Decimal to f64 in PaymentDTO for consistency

+++ rinha_worker/src/dto.rs
@@ -1,9 +1,8 @@
-use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 
 #[derive(Clone, Debug, Deserialize, Serialize)]
 pub struct PaymentDTO {
-    pub amount: Decimal,
+    pub amount: f64,
     #[serde(rename = "correlationId")]
     pub correlation_id: String,
     #[serde(default)]
@@ -13,7 +12,6 @@ pub struct PaymentDTO {
     pub requested_at: String,
 }
 
-#[derive(Debug)]
 pub enum PaymentProcessor {
     Default,
     Fallback,
